requirements.txt
# requirements.txt
flask==2.0.1
flask-cors==3.0.10
streamlit==1.22.0
pandas==1.4.4
beautifulsoup4==4.9.3
nameparser==1.1.1
email_validator==1.1.3
requests==2.26.0
openpyxl==3.0.9
Werkzeug==2.0.1
rich==10.12.0
tqdm==4.62.3
dnspython
email-validator
requests